Added useSignalR tests
