+++ src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Switch } from 'react-router-dom';
 import Chat from './components/chat/Chat';
 import LandingPage from './components/home/LandingPage';
 import ProtectedRoute from './components/ProtectedRoute';
-import { SignalRProvider } from './hooks/useSignalR';
+import SignalRProvider from './components/SignalRProvider';
 
 const App = () => (
 	<BrowserRouter>
